separate json typings.
Some of the types are only relevant to one thing.
